FIX: 1. Some little changes to the Addonsystem. 2. Fixed removeaddon
update you mangosd.conf

git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@1604 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Chat.h
@@ -72,7 +72,7 @@ class ChatHandler
         ChatHandler();
         ~ChatHandler();
 
-        void FillMessageData( WorldPacket *data, WorldSession* session, uint32 type, uint32 language, const char* channelName, const char* message ) const;
+        void FillMessageData( WorldPacket *data, WorldSession* session, uint8 type, uint32 language, const char* channelName, const char* message ) const;
         void FillSystemMessageData( WorldPacket *data, WorldSession* session, const char* message ) const
         {
             FillMessageData( data, session, CHAT_MSG_SYSTEM, LANG_UNIVERSAL, NULL, message );
